requirements.txt
fastapi
sqlalchemy
strawberry-graphql[fastapi]
uvicorn[standard]
bcrypt
pytest
pyjwt